feat(proxy): add proxy print settings bar and extract print hook
